requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pydantic>=2.0.0
tqdm>=4.65.0
pandas>=2.0.0
SimpleITK>=2.3.0
Pillow>=9.5.0
openslide-python>=1.3.0
timm>=0.9.0
vit-pytorch>=0.43.0
einops>=0.6.0
huggingface-hub>=0.16.0
transformers>=4.30.0
pytorch-lightning>=2.0.0
wandb>=0.15.0
x-transformers>=1.0.0
matplotlib>=3.7.0
lifelines>=0.27.0
